requirements.txt
allure_python_commons==2.13.5
Faker==28.4.1
pytest==8.2.2
pytest==8.3.3
Requests==2.32.3
